+++ packages/components/src/components/combobox/combobox.browser.e2e.tsx
@@ -15,6 +15,7 @@ import {
   openClose,
   reflects,
   renders,
+  scalePropagates,
   t9n,
   themed,
   topLayer,
@@ -213,6 +214,21 @@ describe("internal label", () => {
 
 describe("renders", () => {
   renders(() => mount("calcite-combobox"), { display: "block" });
+});
+
+describe("propagates", () => {
+  scalePropagates(
+    (mountOptions) =>
+      mount(
+        <calcite-combobox>
+          <calcite-combobox-item-group>
+            <calcite-combobox-item selected value="item" />
+          </calcite-combobox-item-group>
+        </calcite-combobox>,
+        mountOptions,
+      ),
+    { targetSelector: "calcite-combobox-item, calcite-combobox-item-group, calcite-chip" },
+  );
 });
 
 describe("focusable", () => {